fix: broadcast to all peers
Since we will potentially have a lot more peers than other nodes,
as we are bridging two mutually incompatible versions, the
algorithm for determining when to broadcast a block can mean peers
on the incompatible side won't receive it if there are not many
bridges and they are relying on our node. So it's safer to simply
broadcast the block to everyone all the time.