fix: call original method and recalculate balances
It is necessary to call the original updateVoteBalances method in
order to calculate the difference in vote weight between that and
our method, which uses an otherwise stock Core 3.0 implementation.
This is to accommodate chains with custom logic, such as staking,
by ensuring the staked weight (or any other custom value) is still
applied to the votes, or else we cannot maintain sync since our
votes will not match with the rest of the network.